+++ lib/feature_pack/api/controller.rb
@@ -1,3 +1,7 @@
-# Base controller for API-only features (no view rendering)
-class FeaturePack::API::Controller < ActionController::API
+module FeaturePack
+  module API
+    # Base controller for API-only features (no view rendering)
+    class Controller < ActionController::API
+    end
+  end
 end